backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy==1.26.4
pandas==2.0.3
scikit-learn==1.5.0
tensorflow==2.18.0
joblib==1.3.2
python-dotenv==1.0.0
gunicorn==21.2.0
scipy==1.13.1
openpyxl==3.1.2
requests==2.31.0
apscheduler==3.10.4
python-dateutil==2.9.0.post0
